Separé el proyecto en 2 archivos, 1 para la función main y el otro para las clases, aparte de agregar cambios al spawn de los meteoritos
